dripy/requirements.txt
pynacl
enum